[changelog] create PRs instead of directly pushing
